Move comment accordion to own component
