--- src/components/DetailView/Components/CommentSection/index.ts
@@ -1,1 +1,0 @@
-export * from "./CommentSection";
