+++ src/components/issue-details/Comment/index.ts
@@ -1,0 +1,2 @@
+export * from "./CommentSectionAccordion";
+export * from "./CommentSection";
